requirements.txt
chromadb>=0.4.0
ollama>=0.1.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
pyyaml>=6.0
markdown>=3.5.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
